requirements.txt
streamlit
numpy
pandas
matplotlib
scipy
plotly
yfinance
Pillow
